Move wait_for_source_complete_event() to measurement.py
